+++ Examples/small_quadratic_resolve_example/main.cc
@@ -48,11 +48,11 @@ void init_les( MatrixT& _A, std::vector< double >& _b)
 template<class MatrixT>
 void print_equations( const MatrixT& _B)
 {
-  int m = gmm::mat_nrows( _B);
-  int n = gmm::mat_ncols( _B);
-  for( int i = 0; i < m; ++i)
+  size_t m = gmm::mat_nrows( _B);
+  size_t n = gmm::mat_ncols( _B);
+  for( size_t i = 0; i < m; ++i)
   {
-    for( int j = 0; j < n-1; ++j)
+    for( size_t j = 0; j < n-1; ++j)
     {
       if( _B(i,j) != 0.0)
         std::cout << _B(i,j) << "*x" << j;
